backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
supabase==2.0.3
crewai==0.11.0
langchain==0.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pytest==7.4.3
qdrant-client==1.7.0
neo4j==5.14.0
numpy==1.26.0
openai==1.7.1  # For embeddings generation 